以下の型を修正 ・dp::WindowPositionEventHandler ・dp::WindowSizeEventHandler ・dp::WindowPaintEventHandler 以下の関数の引数の型を修正 ・dp::newWindow()(サイズのみ) ・dp::newWindow()(サイズとdp::WindowFlags) ・dp::newWindow()(サイズと位置 ・dp::newWindow()(サイズと位置とdp::WindowFlags) ・dp::setPosition() ・dp::setSize() ・dp::repaint() ・dp::callPositionEventHandler() ・dp::callSizeEventHandler() ・dp::callPaintEventHandler() ・dp::initizlizeWindowImpl()(サイズのみ) ・dp::initizlizeWindowImpl()(サイズと位置)

+++ impl/inc/dpimpl/common/window/window.h
@@ -23,14 +23,14 @@ namespace dp {
 
     void callPositionEventHandler(
         Window &
-        , Long
-        , Long
+        , Int
+        , Int
     );
 
     void callSizeEventHandler(
         Window &
-        , ULong
-        , ULong
+        , Int
+        , Int
     );
 
     void callBeginPaintEventHandler(
@@ -43,10 +43,10 @@ namespace dp {
 
     void callPaintEventHandler(
         Window &
-        , ULong
-        , ULong
-        , ULong
-        , ULong
+        , Int
+        , Int
+        , Int
+        , Int
     );
 
     struct WindowImpl;
@@ -54,18 +54,18 @@ namespace dp {
     Bool initializeWindowImpl(
         Window &
         , const Utf32 &
-        , ULong
-        , ULong
+        , Int
+        , Int
         , WindowFlags
     );
 
     Bool initializeWindowImpl(
         Window &
         , const Utf32 &
-        , Long
-        , Long
-        , ULong
-        , ULong
+        , Int
+        , Int
+        , Int
+        , Int
         , WindowFlags
     );
